problem: [/] no filtering for future projects #1101
solution: added filter for future projects and verified add new coin procedure
