+++ imports/ui/pages/addCoin/addCoin.scss
@@ -10,3 +10,21 @@
 		width: calc(100%/6) !important;
 	}
 }
+
+#addCurrency {
+  .add-currency-form-container {
+    padding-top: 10px;
+
+    .add-coin-label {
+      margin-left: 15px;
+    }
+    .card{
+      @media only screen and (min-width : 768px) {
+        width: calc( 100vw - 300px);
+      }
+      @media only screen and (max-width : 768px) {
+        width: 90%;
+      }
+    }
+  }
+}
