+++ imports/ui/pages/returnedCurrencies/returnedCurrencies.html
@@ -13,7 +13,7 @@
           <input type="text" class="date-slider span2"/>
         </div>
         <h5 class="card-title mt-2">Security</h5>
-        <div class="form-check">
+        <div class="form-check security-constraints">
           <input type="checkbox" class="form-check-input" id="pow" value="Proof of Work" checked>
           <label class="form-check-label" for="pow">Proof of Work</label>
           <br/>
@@ -22,7 +22,16 @@
           <br/>
           <input type="checkbox" class="form-check-input" id="hybrid" value="Hybrid" checked>
           <label class="form-check-label" for="pos" id="hybrid" >Hybrid</label>
+          <br/>
+          <input type="checkbox" class="form-check-input" id="unknown" value="--Select One--" checked>
+          <label class="form-check-label" for="unknown" id="unknown" >Unknown</label>
         </div>
+        <br/>
+        <hr>
+        <span class="future-projects-option">
+          <input type="checkbox" class="form-check-input" id="future-projects-checkbox" value="Future" checked>
+          <label class="form-check-label" for="future-projects-checkbox">show projects that haven't launched yet</label>
+        </span>
       </div>
       <div class="modal-footer">
         <button type="button" class="btn btn-secondary apply-filter-button" data-dismiss="modal">Apply Filter</button>
