+++ imports/ui/pages/returnedCurrencies/returnedCurrencies.scss
@@ -18,6 +18,16 @@
     }
   }
 
+  .horizontal-separator {
+    border-top: 1px solid #cccccc;
+    width: 100%;
+    display: inline-block;
+  }
+
+  .future-projects-option {
+    padding: 5px 0px 15px 20px;
+  }
+
   .apply-filter-button {
     background-color: #eeeeee;
   }
