+++ package.json
@@ -29,6 +29,7 @@
     "meteor-node-stubs": "~0.2.4",
     "mime-types": "^2.1.17",
     "moment": "^2.20.1",
+    "popper.js": "^1.14.3",
     "recaptcha-v2": "^0.1.3",
     "scrollmagic": "^2.0.5",
     "simpl-schema": "^1.4.2",
